move origin classes to package 'types'
