+++ src/commonMain/kotlin/fr/acinq/lightning/bin/db/payments/IncomingQueries.kt
@@ -19,9 +19,7 @@ package fr.acinq.lightning.bin.db.payments
 import app.cash.sqldelight.coroutines.asFlow
 import app.cash.sqldelight.coroutines.mapToList
 import fr.acinq.bitcoin.ByteVector32
-import fr.acinq.lightning.bin.db.payments.types.IncomingReceivedWithData
-import fr.acinq.lightning.bin.db.payments.types.IncomingReceivedWithTypeVersion
-import fr.acinq.lightning.bin.db.payments.types.mapToDb
+import fr.acinq.lightning.bin.db.payments.types.*
 import fr.acinq.lightning.db.IncomingPayment
 import fr.acinq.phoenix.db.PhoenixDatabase
 import kotlinx.coroutines.Dispatchers
